Ditched the old system of adding in data and now Use the cca api (no more pain)

+++ src/main/kotlin/com/theendercore/npctrader/networking/packet/s2c/TradeMenuS2CPacket.kt
@@ -5,17 +5,17 @@ import net.minecraft.network.Packet
 import net.minecraft.network.PacketByteBuf
 import net.minecraft.network.listener.ClientPlayPacketListener
 
-class TradeMenuS2CPacket(val currency: Int = 69, val trader: Int) : Packet<ClientPlayPacketListener?> {
+class TradeMenuS2CPacket(val currency: Long = 69, val trader: Int) : Packet<ClientPlayPacketListener?> {
 
-    constructor(buf: PacketByteBuf) : this(buf.readInt(), buf.readInt())
+    constructor(buf: PacketByteBuf) : this(buf.readLong(), buf.readInt())
 
     override fun write(buf: PacketByteBuf) {
-        buf.writeInt(currency); buf.writeInt(trader)
+        buf.writeLong(currency); buf.writeInt(trader)
     }
 
     fun toBuf(): PacketByteBuf {
         val buf = PacketByteBuf(Unpooled.buffer())
-        buf.writeInt(currency)
+        buf.writeLong(currency)
         buf.writeInt(trader)
         return buf
     }
